Leave Type with total days used and delete approved leave fix

+++ EmployeeLeaveAPI/Endpoints/ApprovedLeavesEndpoints.cs
@@ -92,6 +92,10 @@ public class ApprovedLeavesEndpoints
             try
             {
                 ApprovedLeave leaveToDelete = await appLeaveRepo.GetByRequestId(requestId);
+                if (leaveToDelete == null)
+                {
+                    return Results.NotFound();
+                }
                 var deletedLeave = await repo.Delete(leaveToDelete.ApprovedLeaveId);
                 return deletedLeave != null ? Results.Ok(deletedLeave) : Results.NotFound();
             }
